+++ src/index.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-import './scss/app.scss'
+import {BrowserRouter as Router,} from 'react-router-dom'
 import App from './App';
+import './scss/app.scss'
 
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <Router>
+        <App/>
+    </Router>
+    ,
+    document.getElementById('root')
 );
