+++ src/pages/Cart.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Cart = () => {
+    return (
+        <div>
+            <h1>Hello Basket</h1>
+        </div>
+    );
+};
+
+export default Cart;
