+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export {default as MainContent} from './MainContent'
+export {default as Cart} from './Cart'
