+++ src/scss/components/_sort.scss
@@ -6,8 +6,14 @@
     display: flex;
     align-items: center;
 
+
     svg {
       margin-right: 8px;
+      transform: rotate(180deg);
+
+      &.rotated{
+        transform: rotate(0deg);
+      }
     }
 
     b {
